+++ radio/src/gui/colorlcd/layouts/layout1x1.cpp
@@ -19,7 +19,7 @@
  */
 
 #include "layout.h"
-#include "layout_factory_implem.h"
+#include "layout_factory_impl.h"
 
 const uint8_t LBM_LAYOUT_1x1[] = {
 #include "mask_layout1x1.lbm"
