+++ radio/src/gui/colorlcd/widgets/timer.cpp
@@ -19,7 +19,7 @@
  */
 
 #include "opentx.h"
-#include "widgets_container_implem.h"
+#include "widgets_container_impl.h"
 
 class TimerWidget: public Widget
 {
